+++ Sources/UI/ShoppingCart/ShoppingCartViewControllerSwiftUI.swift
@@ -35,8 +35,4 @@ open class ShoppingCartViewController: UIHostingController<ShoppingCartView> {
     @MainActor required dynamic public init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
-    open override func viewDidLoad() {
-        super.viewDidLoad()
-    }
 }